+++ src/endpoint/simple-fastapi-container/src/playground/package.json
@@ -10,6 +10,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-markdown": "^9.0.0",
+    "react-router-dom": "^6.19.0",
     "rehype-external-links": "^3.0.0",
     "remark-gfm": "^4.0.0",
     "web-vitals": "^2.1.4"
